--- a/dynparquet/reader.go
+++ b/dynparquet/reader.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 
 	"github.com/parquet-go/parquet-go"
+	"github.com/parquet-go/parquet-go/format"
 )
 
 const (
@@ -18,6 +19,23 @@ type SerializedBuffer struct {
 	f       *parquet.File
 	dynCols map[string][]string
 	fields  []parquet.Field
+
+	indexes []format.ColumnIndex
+	offsets []format.OffsetIndex
+}
+
+type SerializedBufferOption func(*SerializedBuffer)
+
+func WithIndexes(indexes []format.ColumnIndex) SerializedBufferOption {
+	return func(b *SerializedBuffer) {
+		b.indexes = indexes
+	}
+}
+
+func WithOffsets(offsets []format.OffsetIndex) SerializedBufferOption {
+	return func(b *SerializedBuffer) {
+		b.offsets = offsets
+	}
 }
 
 func ReaderFromBytes(buf []byte) (*SerializedBuffer, error) {
@@ -29,7 +47,7 @@ func ReaderFromBytes(buf []byte) (*SerializedBuffer, error) {
 	return NewSerializedBuffer(f)
 }
 
-func NewSerializedBuffer(f *parquet.File) (*SerializedBuffer, error) {
+func NewSerializedBuffer(f *parquet.File, options ...SerializedBufferOption) (*SerializedBuffer, error) {
 	dynColString, found := f.Lookup(DynamicColumnsKey)
 	if !found {
 		return nil, ErrNoDynamicColumns
@@ -40,11 +58,17 @@ func NewSerializedBuffer(f *parquet.File) (*SerializedBuffer, error) {
 		return nil, fmt.Errorf("deserialize dynamic columns metadata %q: %w", dynColString, err)
 	}
 
-	return &SerializedBuffer{
+	sb := &SerializedBuffer{
 		f:       f,
 		dynCols: dynCols,
 		fields:  f.Schema().Fields(),
-	}, nil
+	}
+
+	for _, option := range options {
+		option(sb)
+	}
+
+	return sb, nil
 }
 
 func (b *SerializedBuffer) Reader() *parquet.GenericReader[any] {
@@ -87,13 +111,31 @@ func (b *SerializedBuffer) String() string {
 }
 
 type serializedRowGroup struct {
-	parquet.RowGroup
-	dynCols map[string][]string
-	fields  []parquet.Field
+	indexes  []format.ColumnIndex
+	offsets  []format.OffsetIndex
+	rowGroup parquet.RowGroup
+	dynCols  map[string][]string
+	fields   []parquet.Field
 }
 
 func (b *SerializedBuffer) DynamicRowGroup(i int) DynamicRowGroup {
-	return b.newDynamicRowGroup(b.f.RowGroups()[i])
+	srg := &serializedRowGroup{
+		rowGroup: b.f.RowGroups()[i],
+		dynCols:  b.dynCols,
+		fields:   b.fields,
+	}
+
+	if b.indexes != nil {
+		cols := len(b.f.Schema().Fields())
+		srg.indexes = b.indexes[i*cols : (i*cols)+cols]
+	}
+
+	if b.offsets != nil {
+		cols := len(b.f.Schema().Fields())
+		srg.offsets = b.offsets[i*cols : (i*cols)+cols]
+	}
+
+	return srg
 }
 
 // MultiDynamicRowGroup returns all the row groups wrapped in a single multi
@@ -104,7 +146,7 @@ func (b *SerializedBuffer) MultiDynamicRowGroup() DynamicRowGroup {
 
 func (b *SerializedBuffer) newDynamicRowGroup(rowGroup parquet.RowGroup) DynamicRowGroup {
 	return &serializedRowGroup{
-		RowGroup: rowGroup,
+		rowGroup: rowGroup,
 		dynCols:  b.dynCols,
 		fields:   b.fields,
 	}
@@ -125,3 +167,46 @@ func (g *serializedRowGroup) DynamicRows() DynamicRowReader {
 func (b *SerializedBuffer) DynamicColumns() map[string][]string {
 	return b.dynCols
 }
+
+func (b *serializedRowGroup) NumRows() int64 { return b.rowGroup.NumRows() }
+func (b *serializedRowGroup) ColumnChunks() []parquet.ColumnChunk {
+	chunks := b.rowGroup.ColumnChunks()
+	if b.indexes != nil && b.offsets != nil {
+		cachedColumnChunks := make([]parquet.ColumnChunk, len(chunks))
+		for i, chunk := range chunks {
+			chunks[i] = &cachedColumnChunk{
+				index:  b.indexes[i],
+				offset: b.offsets[i],
+				chunk:  chunk,
+			}
+		}
+
+		return cachedColumnChunks
+	}
+
+	return chunks
+}
+
+func (b *serializedRowGroup) Schema() *parquet.Schema { return b.rowGroup.Schema() }
+func (b *serializedRowGroup) SortingColumns() []parquet.SortingColumn {
+	return b.rowGroup.SortingColumns()
+}
+func (b *serializedRowGroup) Rows() parquet.Rows { return b.rowGroup.Rows() }
+
+type cachedColumnChunk struct {
+	index  format.ColumnIndex
+	offset format.OffsetIndex
+	chunk  parquet.ColumnChunk
+}
+
+func (c *cachedColumnChunk) Type() parquet.Type   { return c.chunk.Type() }
+func (c *cachedColumnChunk) Column() int          { return c.chunk.Column() }
+func (c *cachedColumnChunk) Pages() parquet.Pages { return c.chunk.Pages() }
+func (c *cachedColumnChunk) ColumnIndex() parquet.ColumnIndex {
+	return parquet.NewColumnIndex(c.chunk.Type().Kind(), &c.index)
+}
+func (c *cachedColumnChunk) OffsetIndex() parquet.OffsetIndex {
+	return (*parquet.FileOffsetIndex)(&c.offset)
+}
+func (c *cachedColumnChunk) BloomFilter() parquet.BloomFilter { return c.chunk.BloomFilter() }
+func (c *cachedColumnChunk) NumValues() int64                 { return c.chunk.NumValues() }